+++ doc/changelog.rst
@@ -14,6 +14,7 @@ Version 0.7.6
 * Run samples can now be integer (not only float).
 * Don't round samples to 1 nanosecond anymore: with a large number of loops
   (ex: 2^24), rounding reduces the accuracy.
+* The benchmark calibration is now done by the first worker process
 
 Version 0.7.5 (2016-09-01)
 --------------------------